vibe-dependency-app/backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
click>=7.1.2,<8.0.0
itsdangerous>=2.0.0,<2.1.0
Jinja2>=3.0.0,<3.1.0
google-cloud-storage==2.5.0
numpy<2.0.0
networkx==2.7.1
matplotlib==3.5.1
flask-cors==3.0.10
python-dotenv==0.20.0
pygraphviz==1.9
plotly>=5.10.0
pandas>=1.3.0 